+++ src/components/skip-link/index.ts
@@ -1,0 +1,8 @@
+export {
+  SkipLink,
+  SkipLinkRoot,
+  SkipLinkTarget,
+  type SkipLinkRootProps,
+  type SkipLinkTargetProps,
+} from "./SkipLink.js";
+